BEP/Project/requirements.txt
django==2.1.5
django-webpack-loader==0.6.0
django-rest-framework==0.1.0
djangorestframework-simplejwt==4.3.0
django-cors-headers==2.5.2
django-filter==2.1.0
mysql==0.0.2
datetime==4.3
pillow==6.0.0

pylint-django==2.0.6
